% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{what_time}
\alias{what_time}
\title{Current time}
\usage{
what_time(language = "fr")
}
\arguments{
\item{language}{Language either "fr" for French or "en" for English.}
}
\value{
A character string with the time
}
\description{
Current time
}
\examples{
what_time()
}
