% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{what_time}
\alias{what_time}
\title{Give the time}
\usage{
what_time(language = "fr")
}
\arguments{
\item{language}{Language (character)}
}
\value{
Text (character) with the time
}
\description{
Give the time
}
\examples{
what_time()
}
